refactor: move test  helpers file
