+++ bin/test.ts
@@ -3,7 +3,7 @@ import { apiClient } from '@japa/api-client'
 import { fileSystem } from '@japa/file-system'
 import { configure, processCLIArgs, run } from '@japa/runner'
 
-import { BASE_URL } from '../test_helpers/setup.js'
+import { BASE_URL } from '../test/helpers.js'
 
 processCLIArgs(process.argv.slice(2))
 configure({
