+++ test/configure.spec.ts
@@ -2,7 +2,7 @@ import { test } from '@japa/runner'
 import { IgnitorFactory } from '@adonisjs/core/factories'
 import Configure from '@adonisjs/core/commands/configure'
 
-import { BASE_URL } from '../test_helpers/setup.js'
+import { BASE_URL } from './helpers.js'
 
 test.group('Configure package', () => {
   test('create config file and add provider', async ({ assert, fs }) => {
